src/requirements.txt
blinker==1.4
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
more-itertools==8.10.0
numpy==1.26.4
oauthlib==3.2.0
opencv-python==4.9.0.80
pdf2image==1.17.0
pillow==10.2.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyocr==0.8.5
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
unattended-upgrades==0.1
wadllib==1.3.6
websockets==12.0
zipp==1.0.0
